+++ signer/src/extended_key.rs
@@ -7,65 +7,10 @@ use sha2::Sha512;
 use std::convert::TryFrom;
 use std::fmt;
 use zeroize::Zeroize;
-
-const HARDENED_BIT: u32 = 1 << 31;
-
-pub struct BIP44Path(pub [u32; 5]);
-
-/// Bip44Path
-///
-/// Implementation of the BIP44 standard for derivation path.
-///
-impl BIP44Path {
-    pub fn from_slice(path: &[u32]) -> Result<BIP44Path, SignerError> {
-        let mut path_array: [u32; 5] = Default::default();
-        if path.len() != 5 {
-            return Err(SignerError::GenericString(
-                "Invalid length for path".to_string(),
-            ));
-        };
-
-        path_array.copy_from_slice(path);
-
-        Ok(BIP44Path(path_array))
-    }
-
-    pub fn from_string(path: &str) -> Result<BIP44Path, SignerError> {
-        let mut path = path.split('/');
-
-        if path.next() != Some("m") {
-            return Err(SignerError::GenericString(
-                "Path should start with `m`".to_string(),
-            ));
-        };
-
-        let result = path
-            .map(|index| {
-                let (index_to_parse, mask) = if index.ends_with('\'') {
-                    // Remove the last character and harden index
-                    (&index[..index.len() - 1], HARDENED_BIT)
-                } else {
-                    (index, 0)
-                };
-
-                // FIX ME
-                let child_index = index_to_parse.parse::<u32>()?;
-
-                Ok(child_index | mask)
-            })
-            .collect::<Result<Vec<u32>, std::num::ParseIntError>>()?;
-
-        let bip44_path = BIP44Path::from_slice(&result)?;
-
-        Ok(bip44_path)
-    }
-
-    pub fn is_testnet(&self) -> bool {
-        return self.0[1] == (1 | HARDENED_BIT);
-    }
-}
+use bip44::BIP44Path;
 
 const HMAC_SEED: &[u8; 12] = b"Bitcoin seed";
+const HARDENED_BIT: u32 = 1 << 31;
 
 #[derive(Zeroize, Debug)]
 #[zeroize(drop)]
@@ -176,7 +121,8 @@ impl ExtendedSecretKey {
 
 #[cfg(test)]
 mod tests {
-    use crate::bip44::{BIP44Path, ExtendedSecretKey};
+    use crate::extended_key::ExtendedSecretKey;
+    use bip44::BIP44Path;
     use bip39::{Language, Mnemonic, Seed};
     use hex::encode;
     use std::convert::TryFrom;
